move genfile to astbasic
